docs: Update to README with webserver info
This commit contains changes that were made to the README file. More
specifically, these are the final changes that will be made to the
file to reflect all the crucial information of the Ubuntu Web Server
that I set up using Amazon Lightsail.

The following information on the server was added:
 - Access information (IP address, SSH Port, and URL)
 - Tools installed (Server software and Python modules)
 - List of resources for setting up