docs: Update to README with configuration summary
This commit contains changes to the README file. A summary of the
configurations of the Ubuntu web server was added to provide more
concise details about how the server was configured.

Prior to this commit, the README lacked important configuration
details about the Ubuntu web server. Provide this information gives
reader better details about the web server itself.